feature: warning when exceptionIds are unused.
When the list of exceptions in the .nsprc file grows it is nice that we
show that exceptions are no longer needed. This way developers have
an easy time pruning the exceptions.

By comparing the exceptionIds with the actual found vulnerabilities
we can check if the exceptionIds were actually used or not. Then we
can show an info message, containing all exceptionIds which are not
used.

Had to split up the common `getVulnerabilities` into two functions:
`getRawVulnerabilities` and `filterExceptions`.

The `getRawVulnerabilities` now simply returns the vulnerabilities as
reported by `npm audit` normalized, but without the exceptionIds
excluded. This way we can calculate the unusedExceptionIds

Had to pass along the unusedExceptionIds's from the `audit` function to the
`auditLog` function, and then finally to the `handleFinish` function
so it can be logged there.

Also fixed some spelling mistakes.