Allow vulnerabilities to be identifed by URL
Unfortunately, for mysterious reasons, the ids returned from the NPM
audit API have been changing recently, which makes identifying the
vulnerabilities difficult.

Even more unfortunately, when NPM changed the output format for
`npm audit --json` in npm v7, they removed many of the other useful
identifiers (`cves`, `cwe`, `github_advisory_id`). The only thing left
is the URL.

This patch allows vulnerabilities in the `.nsprc` file to be identified
by the URL instead of the sadly unstable id. This doesn't seem like a
really great option to me, since all the URLs changed from npmjs.com to
github.com a few months ago, and surely they'll change again. But it's
the closest thing we have to a stable id at the moment, so we'll go with
that.

See #60